Move select tile group
