+++ forui/lib/widgets/select_group.dart
@@ -1,6 +1,6 @@
 /// {@category Widgets}
 ///
-/// A group of form fields that represent a single selection.
+/// A group of form fields that allow users to make a selection from a set of options.
 ///
 /// See https://forui.dev/docs/form/select-group for working examples.
 library forui.widgets.select_group;
