+++ forui/lib/widgets/select_tile.dart
@@ -1,0 +1,9 @@
+/// {@category Widgets}
+///
+/// A group of tiles that allow users to make a selection from a set of options.
+///
+/// See https://forui.dev/tile/select-tile-group for working examples.
+library forui.widgets.select_tile;
+
+export '../src/widgets/select_tile_group/select_tile.dart' hide FSelectTileData;
+export '../src/widgets/select_tile_group/select_tile_group.dart';
