+++ forui/lib/widgets/tile.dart
@@ -1,12 +1,10 @@
 /// {@category Widgets}
 ///
-/// The numerous tile variants.
+/// A tile group that is typically used to group related information together.
 ///
 /// See https://forui.dev/tile for working examples.
 library forui.widgets.tile;
 
-export '../src/widgets/tile/select_tile.dart' hide FSelectTileData;
-export '../src/widgets/tile/select_tile_group.dart';
 export '../src/widgets/tile/tile.dart' hide extractTile;
 export '../src/widgets/tile/tile_content.dart' hide FTileContent;
 export '../src/widgets/tile/tile_group.dart' hide extractTileGroup;
