+++ forui/test/golden/select-tile-group/group/empty.png
