+++ forui/test/golden/select-tile-group/group/override-state.png
