+++ forui/test/golden/select-tile-group/group/override-style.png
