+++ forui/test/golden/select-tile-group/group/single.png
