+++ forui/test/golden/select-tile-group/group/zinc-dark/disabled/FTileDivider.full-bottom.png
