+++ forui/test/golden/select-tile-group/group/zinc-dark/disabled/FTileDivider.none-bottom.png
