+++ forui/test/golden/select-tile-group/group/zinc-dark/enabled/FTileDivider.indented.png
