+++ forui/test/golden/select-tile-group/group/zinc-light/disabled/FTileDivider.full-bottom.png
