+++ forui/test/golden/select-tile-group/group/zinc-light/disabled/FTileDivider.none-top.png
