+++ forui/test/golden/select-tile-group/group/zinc-light/enabled/FTileDivider.full.png
