+++ forui/test/golden/select-tile-group/group/zinc-light/enabled/FTileDivider.indented.png
