+++ forui/test/golden/select-tile-group/group/zinc-light/enabled/FTileDivider.none.png
