+++ forui/test/golden/select-tile-group/group/zinc-light/error.png
