+++ forui/test/golden/select-tile-group/group/zinc-light/hovered/FTileDivider.full-bottom.png
