+++ forui/test/golden/select-tile-group/group/zinc-light/hovered/FTileDivider.full-top.png
