+++ forui/test/golden/select-tile-group/group/zinc-light/hovered/FTileDivider.none-bottom.png
