+++ forui/test/golden/select-tile-group/tile/checked-zinc-dark.png
