+++ forui/test/golden/select-tile-group/tile/disabled-zinc-light.png
