+++ forui/test/golden/select-tile-group/tile/enabled-zinc-dark.png
