+++ forui/test/golden/select-tile-group/tile/enabled-zinc-light.png
