+++ forui/test/golden/select-tile-group/tile/hovered-zinc-dark.png
