+++ forui/test/golden/select-tile-group/tile/hovered-zinc-light.png
