+++ forui/test/golden/select-tile-group/tile/suffix-checked-zinc-light.png
