+++ forui/test/golden/select-tile-group/tile/suffix-unchecked-zinc-dark.png
